Add (prepend) context to message directly
When logging large amounts of data out the extra context is often
truncated from view. To be able to tell which function is making a
particular call in the logs we need to prepend this data in a more
visible manner.

Problems are occuring when trying to subscribe - it appears the request
is invalid somehow, but cannot figure out how from existing logs. More
detail is needed... such as context.