Update subscriber data to include required field
The Campaign Monitor API will issue an HTTP 400 with a message to set
the ConsentToTrack value if it is not present. This in turn can result
in an HTTP 500 due to an uncaught CMError on the requesting site.

We can also short-circuit invalid API calls by detecting invalid IDs
when asked to load by ID instead of waiting for a network round-trip and
an unclear reason for the error (simply a 404).